requisicao de login funcionando. Detalhes finais para cadastro de prestdor.

+++ ConectaProApp/Models/Plano.cs
@@ -9,7 +9,7 @@ namespace ConectaProApp.Models
 {
     public class Plano
     {
-        public int IdPlano { get; set; }
+        public long IdPlano { get; set; }
         public string DescPlano { get; set; }
         public float ValorFixoPlano { get; set; }
         public float PercentualPlano { get; set; }
